requirements.txt
discord.py>=2.4.0
google-generativeai==0.3.2
python-dateutil==2.8.2
aiohttp>=3.9.0
python-dotenv==1.0.0
flask==3.0.0
